requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jinja2==3.1.2 